move files, add iterations to random tests, remove cse function
